Load a json file that defines all the commands available, and their operation types
I deem "not harmful" to have an init() function that loads the commands in memory as an easier way to maintain those commands in the program, and being able to read them from other scripts (such as redis-commands.py). Whenever I migrate the Python script into Go it will make sense to move all this information into a hardcoded struct and remove the init() function altogether.